types reorg and export
