+++ packages/visualizations/src/index.ts
@@ -10,3 +10,4 @@ export * from './components/Chart/types';
 export * from './components/KpiCard/types';
 export * from './components/Map/types';
 export * from './components/MarkdownText/types';
+export * from './components/Legend/types';
